Returns working

+++ src/main/kotlin/conditions/equality/IntEquality.kt
@@ -3,6 +3,6 @@ package conditions.equality
 import attributes.HasAttributes
 import conditions.Condition
 
-class IntEquality(val attribute: HasAttributes.IntAttribute, val value: Int): Condition{
+class IntEquality(val attribute: HasAttributes.IntAttribute, val value: Long): Condition{
     override fun toString() = "${attribute.getString()} = $value"
 }
